Se agregan relaciones

+++ src/main/java/com/cursoSpring/persistencia/entity/ComprasProductosPk.java
@@ -2,29 +2,32 @@ package com.cursoSpring.persistencia.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 
 import java.io.Serializable;
 
 @Embeddable
 public class ComprasProductosPk implements Serializable {
     @Column(name="id_compra")
-    private Long idCompra;
+    private Integer idCompra;
     @Column(name="id_productox")
-    private Long idProducto;
+    private Integer idProducto;
 
-    public Long getIdCompra() {
+
+    public Integer getIdCompra() {
         return idCompra;
     }
 
-    public void setIdCompra(Long idCompra) {
+    public void setIdCompra(Integer idCompra) {
         this.idCompra = idCompra;
     }
 
-    public Long getIdProducto() {
+    public Integer getIdProducto() {
         return idProducto;
     }
 
-    public void setIdProducto(Long idProducto) {
+    public void setIdProducto(Integer idProducto) {
         this.idProducto = idProducto;
     }
 }
